nettoyage_express_ancien/netexpress_final_v5/requirements/dev.txt
-r base.txt
pytest>=8.3
pytest-django>=4.8

# Ajout 2025 : ReportLab et Jazzmin ne sont pas installés par défaut dans
# cet environnement afin de réduire les dépendances et d’éviter les
# erreurs liées à l’absence de modules système.  Si vous souhaitez
# générer des PDF ou améliorer le thème de l’administration, installez
# manuellement les paquets ci‑dessous :
reportlab>=3.6
django-jazzmin>=2.6